+++ demo/double_click_counters/click_counter_store.js
@@ -1,0 +1,9 @@
+var ClickCounterStore = Cape.createDataStoreClass({
+  init: function() {
+    this.counter = 0;
+  },
+  increment: function() {
+    this.counter++;
+    this.propagate();
+  }
+});