+++ demo/double_click_counters/modular_click_counter.js
@@ -1,0 +1,10 @@
+"use strict";
+
+var ModularClickCounter = Cape.createComponentClass({
+  render: function(m) {
+    m.div(String(this.ds.counter), {
+      class: 'counter',
+      onclick: function(e) { this.ds.increment() }
+    })
+  }
+});